live-session: staged pipeline + downsample-once (+ inline/fused) + native upsample
Stage the extraction pipeline and add a downsample-once workflow with optional
re-upsampling of results back to native resolution.

Pipeline (cnmfe/pipeline.py):
- Carve fit() into composable fit_extract() + evaluate(); fit() is now a thin
  wrapper (fit_mc -> fit_extract -> evaluate). Behaviour preserved.
- CNMFeParams.downscaled(ssub, tsub) rescales unit-bearing fields.
- CNMFe.upsample_to_native(): non-destructive new model with A (bilinear) and
  C/YrA/C_raw (linear) interpolated to native grid/rate; S kept at ds rate,
  background/shifts dropped (inspection/overlay only, not re-fitting).

Downsampling (single write, three entry points):
- cnmfe/downsample.py: streaming block-mean downsample_movie (+ ds_meta.json),
  and upsample_footprints / upsample_traces helpers.
- concat_avis_to_zarr: inline ssub/tsub binning while decoding.
- avi_mc fused AVI->MC: ssub/tsub binning before MC (fit_mc_from_avis), so a
  downsampled run still writes only mc.zarr.

Fixes:
- high_pass_filter_space: int() the cv2 ksize so a float gSig_filt (from
  downscaled params) no longer raises. Identical for integer inputs.

CLIs: run_preprocess / run_mc / run_extract (--ds-meta) / run_evaluate.

Tests: test_stage_split, test_downsample (downsample + downscaled + upsample +
round-trip), plus inline/fused downsampling cases in test_concat_avis /
test_avi_mc. Full suite: 186 passed, 1 xfailed.

(live_runs/run_session.ipynb intentionally not committed — actively in use.)

+++ cnmfe/motion_correction.py
@@ -30,7 +30,10 @@ def high_pass_filter_space(img_orig, gSig_filt):
     if np.isscalar(gSig_filt):
         gSig_filt = [gSig_filt, gSig_filt]
 
-    ksize = tuple([(3 * i) // 2 * 2 + 1 for i in gSig_filt])
+    # ksize must be an int for cv2 (gSig_filt may be a float, e.g. after
+    # CNMFeParams.downscaled scales it by ssub). For integer gSig_filt this
+    # is identical to the historical expression.
+    ksize = tuple([int((3 * i) // 2 * 2 + 1) for i in gSig_filt])
 
     ker = cv2.getGaussianKernel(ksize[0], gSig_filt[0])
     ker2D = ker.dot(ker.T)
